hackthon-backend/requirements.txt
# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# AI Provider SDKs (nur der gewählte wird pro Request genutzt)
openai>=1.55.0
anthropic>=0.39.0
google-genai>=1.0.0

# Config & Env
pydantic-settings>=2.6.0
